Norm completed. add file for checkinf file name after redirection
